+++ ICD.Profound.ConnectPRO/Themes/TouchDisplayInterface/Presenters/Notifications/IncomingCallPresenter.cs
@@ -175,8 +175,8 @@ namespace ICD.Profound.ConnectPRO.Themes.TouchDisplayInterface.Presenters.Notifi
 			if (room == null)
 				return;
 
-			room.OnIncomingCallAnswered += RoomOnIncomingCallAnswered;
-			room.OnIncomingCallRejected += RoomOnIncomingCallRejected;
+			room.Dialing.OnIncomingCallAnswered += RoomOnIncomingCallAnswered;
+			room.Dialing.OnIncomingCallRejected += RoomOnIncomingCallRejected;
 
 			if (room.ConferenceManager == null)
 				return;
@@ -196,8 +196,8 @@ namespace ICD.Profound.ConnectPRO.Themes.TouchDisplayInterface.Presenters.Notifi
 			if (room == null)
 				return;
 
-			room.OnIncomingCallAnswered -= RoomOnIncomingCallAnswered;
-			room.OnIncomingCallRejected -= RoomOnIncomingCallRejected;
+			room.Dialing.OnIncomingCallAnswered -= RoomOnIncomingCallAnswered;
+			room.Dialing.OnIncomingCallRejected -= RoomOnIncomingCallRejected;
 
 			if (room.ConferenceManager == null)
 				return;
@@ -341,7 +341,7 @@ namespace ICD.Profound.ConnectPRO.Themes.TouchDisplayInterface.Presenters.Notifi
 			IIncomingCall call = GetFirstIncomingCall();
 
 			if (call != null && Room != null)
-				Room.RejectIncomingCall(call);
+				Room.Dialing.RejectIncomingCall(call);
 
 			ShowView(false);
 		}
@@ -360,7 +360,7 @@ namespace ICD.Profound.ConnectPRO.Themes.TouchDisplayInterface.Presenters.Notifi
 					: m_IncomingCallsSection.Execute(() => m_IncomingCalls.GetDefault(call));
 
 			if (call != null && Room != null)
-				Room.AnswerIncomingCall(control, call);
+				Room.Dialing.AnswerIncomingCall(control, call);
 
 			ShowView(false);
 		}